Overhaul of access system and macro importing into templates. Also bugfixes.

+++ resource.py
@@ -162,74 +162,80 @@ class RacModelConverter(ModelConverter):
     return f.TextAreaField(**kwargs)
 
 
-# Checks if user is logged in and authorized
-class ResourceSecurityPolicy:
+class Authorization:
 
-  def __init__(self, public_ops=["view", "list"]):
-    self.public_ops = public_ops
-
-  def is_public_op(self, op):
-    return op in self.public_ops
-
-  def is_authorized(self, user, op, instance):
-    authorized = user is not None or self.is_public_op(op)
-    if not authorized:
-      logger.info("Anonymous user not authorized to do '%s'", op)
-    return authorized
-
-  # Checks if user itself has access rights
-  def is_allowed_user(self, user, op, instance):
-    return True
-
-  def is_allowed_public(self, op, instance):
-    return ResourceSecurityPolicy.is_public_instance(instance) and self.is_public_op(op)
-
-  def is_public_or_user_allowed(self, user, op, instance):
-    if user:
-      return self.is_allowed_user(user, op, instance)
+  def __init__(self, is_authorized, message='', only_fields=None, error_code=403):
+    self.is_authorized = is_authorized
+    self.message = message
+    self.error_code = error_code
+    # Privleged means that this authorization would not apply to the public
+    # or a normal user. E.g. a user can only edit their own profile (privilege),
+    # or an admin can see other people's orders
+    self.only_fields = only_fields
+    if is_authorized:
+      logger.debug("Authorized: %s" % message)
     else:
-      return self.is_allowed_public(op, instance)
+      logger.info("UNAUTHORIZED: %s" % message)
 
-  @staticmethod
-  def is_public_instance(instance):
-    try:
-      return instance.is_public() if instance else True
-    except AttributeError:
-      return True
+  def is_privileged(self):
+    return (self.error_code == 403 and self.is_authorized)
 
-  def validate_operation(self, op, instance=None):
-    if not is_allowed_access(g.user):
-      raise ResourceError(403)
-    if not self.is_authorized(g.user, op, instance):
-      raise ResourceError(401)
-    if not self.is_public_or_user_allowed(g.user, op, instance):
-      raise ResourceError(403)
+  def __nonzero__(self):
+    return self.is_authorized
 
+# Checks if user is logged in and authorized
+class ResourceAccessPolicy:
+  model_class = None
+  levels = ['public', 'user', 'private', 'admin']
+  
+  def __init__(self, ops_levels=None, user_field='user'):
+    if not ops_levels:
+      self.ops_levels = {
+        'view': 'public',
+        'list': 'public',
+        '_default': 'admin'
+      }
+    else:
+      self.ops_levels = ops_levels
 
-class AdminWriteSecurityPolicy(ResourceSecurityPolicy):
-  def is_allowed_user(self, user, op, instance):
-    return user.admin
+  def authorize(self, op, instance=None):
+    if op not in self.ops_levels:
+      level = self.ops_levels['_default']
+    else:
+      level = self.ops_levels[op]
+    msg = '%s requires a logged in user' % op
+    if level=='public':
+      return Authorization(True,'%s is a publicly allowed operation' % op)
+    elif level=='user':
+      if g.user:
+        return Authorization(True, msg)
+      else:
+        return Authorization(False, msg, error_code=401) # Denoted that the user should log in first
+    elif level=='private':
+      if not instance:
+        return Authorization(False, 'Error: Cannot apply private access without an instance')
+      instance_user = getattr(instance, self.user_field, None)
+      if not instance_user:
+        return Authorization(False, 'Error: Cannot identify user (field %s) which instance %s belongs to' % (self.user_field, instance))
+      elif not g.user:
+        return Authorization(False, msg, error_code=401) # Denotes that the user should log in first
+      elif not g.user == instance_user:
+        return Authorization(False, '%s is a private operation which requires the owner to be logged in' % op)
+      else:
+        return Authorization(True, '%s have access to do private operation %s on instance %s' % (instance_user, op, instance))
+    elif level=='admin':
+      if not g.user:
+        return Authorization(False, msg, error_code=401) # Denotes that the user should log in first
+      elif not g.user.admin:
+        return Authorization(False, 'Need to be logged in with admin access')
+      elif g.user:
+        return Authorization(True, '%s is an admin' % g.user)
+    return Authorization(False, 'error', 'This is catch all denied authorization, should not be here')
 
-
-
-class ResourcePrivacyPolicy:
-
-  def __init__(self, security_policy, op):
-    self.op = op
-    self.security_policy = security_policy
-
-  # True if visible due to current user, such as being admin
-  def is_private(self, instance):
-    return g.user \
-            and self.security_policy.is_allowed_user(g.user, self.op, instance) \
-            and not self.security_policy.is_allowed_public(self.op, instance)
-
-
-
-class ResourceAccessStrategy:
+class ResourceRoutingStrategy:
   def __init__(self, model_class, plural_name, id_field='id', form_class=None, 
     parent_strategy=None, parent_reference_field=None, short_url=False, 
-    list_filters=None, use_subdomain=False, security_policy=None):
+    list_filters=None, use_subdomain=False, access_policy=None):
     if use_subdomain and parent_strategy:
       raise ValueError("A subdomain-identified resource cannot have parents")
     self.form_class = form_class if form_class else model_form(model_class, base_class=RacBaseForm, converter=RacModelConverter())
@@ -253,7 +259,8 @@ class ResourceAccessStrategy:
     self.default_list_filters = list_filters
     # The field name pointing to a parent resource for this resource, e.g. article.world
     self.parent_reference_field = self.parent.resource_name if (self.parent and not parent_reference_field) else None
-    self.security = security_policy if security_policy else ResourceSecurityPolicy()
+    self.access = access_policy if access_policy else ResourceAccessPolicy()
+    self.access.model_class = self.model_class
 
   def get_url_path(self, part, op=None):
     parent_url = ('/' if (self.parent is None) else self.parent.url_item(None))
@@ -328,17 +335,8 @@ class ResourceAccessStrategy:
   def endpoint_name(self, suffix):
     return self.resource_name + '_' + suffix
 
-  def get_visibility(self, op):
-    return ResourcePrivacyPolicy(self.security, op)
-
-  def check_operation_any(self, op):
-    self.security.validate_operation(op, None)
-
-  def check_operation_on(self, op, instance):
-    if not instance:
-      raise ResourceError(500)
-    self.security.validate_operation(op, instance)
-
+  def authorize(self, op, instance=None):
+    return self.access.authorize(op, instance)
 
 
 class ResourceError(Exception):
@@ -400,7 +398,11 @@ class ResourceHandler(View):
     app.add_url_rule(st.url_item(), subdomain=st.subdomain_part, methods=['PUT', 'POST'], view_func=cls.as_view(st.endpoint_name('replace'), st))
     app.add_url_rule(st.url_item(), subdomain=st.subdomain_part, methods=['PATCH', 'POST'], view_func=cls.as_view(st.endpoint_name('edit'), st))
     app.add_url_rule(st.url_item(), subdomain=st.subdomain_part, methods=['DELETE', 'POST'], view_func=cls.as_view(st.endpoint_name('delete'), st))
+    
+    if current_app:
+      current_app.access_policy[st.resource_name] = st.access
 
+    # print "in register url %s" % app.app
     # /<resource>/[_,view,edit] -> GET:fablr.co/helmgast/, GET:fablr.co/helmgast/view, GET|POST:fablr.co/helmgast/edit
     # /resource/[list,new] -> GET:/world/list, GET:/world/new
     # <resource>.host/[_,view,edit] -> -> GET:helmgast.fablr.co, GET:helmgast.fablr.co/view, GET|POST:helmgast.fablr.co/edit
@@ -450,15 +452,13 @@ class ResourceHandler(View):
           return redirect(url_for('auth.login', next=request.path))
 
       elif err.status_code == 403: # forbidden
-        r['op'] = self.get_post_pairs[r['op']] if r['op'] in self.get_post_pairs else r['op']  # change the effective op
-        r['template'] = self.strategy.item_template()
-        # r[self.strategy.resource_name + '_form'] = form
-        # if json, return json instead of render
+        # r['op'] = self.get_post_pairs[r['op']] if r['op'] in self.get_post_pairs else r['op']  # change the effective op
+        # r['template'] = self.strategy.item_template()
         if r['out'] == 'json':
           return self._return_json(r, err)
-        elif 'template' in r:
-          flash(err.message,'warning')
-          return render_template(r['template'], **r), 403
+        # elif 'template' in r:
+        #   flash(err.message,'warning')
+        #   return render_template(r['template'], **r), 403
         else:
           return err.message, 403
 
@@ -530,7 +530,6 @@ class ResourceHandler(View):
     r['out'] = request.args.get('out','html') # default to HTML
     if 'next' in request.args:
       r['next'] = request.args['next']
-    r['visibility'] = self.strategy.get_visibility(r['op'])
     return r
 
   def _query_url_components(self, r, **kwargs):
@@ -543,13 +542,22 @@ class ResourceHandler(View):
 
   def view(self, r):
     item = r['item']
-    self.strategy.check_operation_on(r['op'], item)
+    auth = self.strategy.authorize(r['op'], item)
+    r['auth'] = auth
+    if not auth:
+      raise ResourceError(auth.error_code, r, message=auth.message)
+    
     r['template'] = self.strategy.item_template()
     return r
 
   def form_edit(self, r):
     item = r['item']
-    self.strategy.check_operation_on(r['op'], item)
+    
+    auth = self.strategy.authorize(r['op'], item)
+    r['auth'] = auth
+    if not auth:
+      raise ResourceError(auth.error_code, r, message=auth.message)
+
     form = self.form_class(obj=item, **r.get('prefill',{}))
     form.action_url = url_for('.' + self.strategy.endpoint_name('edit'), op='edit', **r['url_args'])
     r[self.strategy.resource_name + '_form'] = form
@@ -558,7 +566,11 @@ class ResourceHandler(View):
     return r
 
   def form_new(self, r):
-    self.strategy.check_operation_any(r['op'])
+    auth = self.strategy.authorize(r['op'])
+    r['auth'] = auth
+    if not auth:
+      raise ResourceError(auth.error_code, r, message=auth.message)
+    
     form = self.form_class(request.args, obj=None, **r.get('prefill',{}))
     form.action_url = url_for('.' + self.strategy.endpoint_name('new'), **r['url_args'])
     r[self.strategy.resource_name + '_form'] = form
@@ -567,7 +579,11 @@ class ResourceHandler(View):
     return r
 
   def list(self, r):
-    self.strategy.check_operation_any(r['op'])
+    auth = self.strategy.authorize(r['op'])
+    r['auth'] = auth
+    if not auth:
+      raise ResourceError(auth.error_code, r, message=auth.message)
+    
     listquery = self.strategy.query_list(request.args).filter(**r.get('filter',{}))
     if r.get('parents'):
       # TODO if the name of the parent resource is different than the reference field name 
@@ -587,7 +603,11 @@ class ResourceHandler(View):
     return r
 
   def new(self, r):
-    self.strategy.check_operation_any(r['op'])
+    auth = self.strategy.authorize(r['op'])
+    r['auth'] = auth
+    if not auth:
+      raise ResourceError(auth.error_code, r, message=auth.message)
+
     form = self.form_class(request.form, obj=None)
     if not form.validate():
       r['form'] = form
@@ -602,7 +622,11 @@ class ResourceHandler(View):
 
   def edit(self, r):
     item = r['item']
-    self.strategy.check_operation_on(r['op'], item)
+    auth = self.strategy.authorize(r['op'], item)
+    r['auth'] = auth
+    if not auth:
+      raise ResourceError(auth.error_code, r, message=auth.message)
+
     form = self.form_class(request.form, obj=item)
     if not form.validate():
       r['form'] = form
@@ -619,7 +643,11 @@ class ResourceHandler(View):
 
   def replace(self, r):
     item = r['item']
-    self.strategy.check_operation_on(r['op'], item)
+    auth = self.strategy.authorize(r['op'], item)
+    r['auth'] = auth
+    if not auth:
+      raise ResourceError(auth.error_code, r, message=auth.message)
+
     form = self.form_class(request.form, obj=item)
     if not form.validate():
       r['form'] = form
@@ -633,13 +661,17 @@ class ResourceHandler(View):
 
   def delete(self, r):
     item = r['item']
-    self.strategy.check_operation_on(r['op'], item)
+    auth = self.strategy.authorize(r['op'], item)
+    r['auth'] = auth
+    if not auth:
+      raise ResourceError(auth.error_code, r, message=auth.message)
+
     if not 'next' in r:
       if 'parents' in r:
         r['next'] = url_for('.' + self.strategy.endpoint_name('list'), **self.strategy.parent.all_view_args(getattr(item, self.strategy.parent_reference_field)))
       else:
         r['next'] = url_for('.' + self.strategy.endpoint_name('list'))
-    self.strategy.check_operation_on(r['op'], item)
+    self.strategy.authorize(r['op'], item)
     logger.info("Delete on %s with id %s", self.strategy.resource_name, item.id)
     item.delete()
     return r
